% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{basename}
\alias{basename}
\title{File name without extension name}
\usage{
basename(path)
}
\arguments{
\item{path}{File path string vector.}
}
\description{
Get the file name of a given file path without extension name.
}
