% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{binarySearch.list}
\alias{binarySearch.list}
\title{Binary search on a \code{list}}
\usage{
binarySearch.list(list, find, key, compares = function(a, b) a - b)
}
\description{
Binary search on a \code{list}
}
\seealso{
\code{\link{binarySearch.impl.generic}}
}
