% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{binarySearch}
\alias{binarySearch}
\title{Perform binary search on a sortted sequence.}
\usage{
binarySearch(source, find, key, compares = function(a, b) a - b)
}
\arguments{
\item{source}{The input sequence, this input source sequence must in sorted in
asc or desc order.}

\item{find}{Target key for find the target element in the input source sequence.}

\item{key}{A lambda function that describ how to abstract the key indexer value
from the elements of the input source sequence.}

\item{compares}{The comparision method that using for find target element by compares
the target key and the key indexer values from the source sequence, by
default is comparision based on their numeric values.}
}
\description{
Perform binary search on a sortted sequence.
}
\details{
If there are duplicated items in the \code{source} sequence, then
         you should group the items at first and then perfamence the binary search.
}
