% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{binaryTree.construct.impl}
\alias{binaryTree.construct.impl}
\title{Binary tree build implementation}
\usage{
binaryTree.construct.impl(tree, popX, popName, src, key = NULL,
  key.compares = NULL, debug = TRUE, progressHeader = TRUE)
}
\description{
Binary tree build implementation
}
