% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{binaryTree}
\alias{binaryTree}
\title{This function build a binary tree object}
\usage{
binaryTree(src, key, key.numeric = as.numeric)
}
\arguments{
\item{src}{the input data sequence, which can be \code{data.frame}, \code{list} or
\code{vector}}

\item{key}{An object indexer for get the key value for the elements}

\item{key.numeric}{Convert the key value to numeric value for comparasions.}
}
\value{
A binary tree S4 class object
}
\description{
This function build a binary tree object
}

