% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{cliToken}
\alias{cliToken}
\title{Create cli tokens}
\usage{
cliToken(arguments)
}
\arguments{
\item{arguments}{A character vector}
}
\description{
Add quote char wrapper for commandline argument token
which have whitespace inside.
}
