% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandArgs.R
\name{commandArgs}
\alias{commandArgs}
\title{The commandline helper for Rscript}
\usage{
commandArgs(..., debug = FALSE)
}
\arguments{
\item{...}{a list of commandline arguments for handle
in this commandline helper. the element data in this
list should be a list of key-value pair in formats
like: \code{argument_name = description}. the
\code{description} value shoud be a vector data in
special data format.}

\item{debug}{commandline helper debug options. value can
be a logical switch or a list of debug data for assign
to test of the commandline helper function.}
}
\value{
A list of commandline argument values in key-value
   pair data format. the commandline argument name has been
   trimmed of its prefix characters.
}
\description{
The commandline helper for Rscript
}
\details{
the commandline argument value should be in format
   like \code{[required=TRUE/FALSE, default_value/error_message, description_text]}.
}
