% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{DESCRIPTION}
\alias{DESCRIPTION}
\title{Parse package DESCRIPTION file}
\usage{
DESCRIPTION(packageName)
}
\arguments{
\item{packagename}{The R package name}
}
\description{
This function is an alias of function
    \code{\link{parse.package.description}}
}

