% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{EmptyFile}
\alias{EmptyFile}
\title{Create an empty file}
\usage{
EmptyFile(outfile)
}
\description{
Create an empty file at specific location, if the target file 
  is already exists, then all of the original contents of target file will 
  be removed.
}

