% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure the dir exists}
\usage{
ensure_dir_exists(path)
}
\description{
If the directory path is not exists on
the file system, then this function will create it.
}
