% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Eval}
\alias{Eval}
\title{list module helper}
\usage{
Eval(module, ...)
}
\arguments{
\item{module}{A function to create a \code{list} object a \code{list} object.}
}
\value{
If the module argument is a list, then it will returns directly,
  or the function will be evaluate for produce a list value.
}
\description{
list module helper
}

