% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{File.Open}
\alias{File.Open}
\title{Open a text file for write in text mode}
\usage{
File.Open(file.txt, append = FALSE)
}
\arguments{
\item{file.txt}{Target file path for write in text mode.}

\item{append}{A logical flag to indicate append the data to target file or not?}
}
\description{
æ‰“å¼€ä¸€ä¸ªæ–‡ä»¶å¥æŸ„ï¼Œç„¶åè¿”å›ä¸€ä¸ªå‡½æ•°æŒ‡é’ˆç”¨æ¥ä»¥\code{sprintf}æ ¼å¼åŒ<U+FFFD>
çš„å½¢å¼å‘æ–‡ä»¶ä»¥è¿½åŠ çš„å½¢å¼å†™å…¥æ•°æ®
}

