% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{File.WithExtension}
\alias{File.WithExtension}
\title{Determine path end with a given extension name}
\usage{
File.WithExtension(path, ext)
}
\arguments{
\item{ext}{File extension name without dot.}
}
\description{
Case insensitive
}
