% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{flash_load}
\alias{flash_load}
\title{Load R script in directory}
\usage{
flash_load(dir = getwd())
}
\arguments{
\item{dir}{The script source directory, by default is current workspace.}
}
\value{
Nothing.
}
\description{
Load all of the R script in a given working directory,
   by default is load all script in current directory.
}
