% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{GetType}
\alias{GetType}
\title{å‡½æ•°è¿”å›\code{\link{primitiveTypes}}æšä¸¾ä¹‹ä¸­çš„æŸä¸€ä¸ªç±»å<U+FFFD>}
\usage{
GetType(x)
}
\description{
å‡½æ•°è¿”å›\code{\link{primitiveTypes}}æšä¸¾ä¹‹ä¸­çš„æŸä¸€ä¸ªç±»å<U+FFFD>
}

