% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{global<-}
\alias{global<-}
\title{Set variable in global}
\usage{
global(name) <- value
}
\value{
\code{NULL}
}
\description{
Set variables into the global environment.
}
\examples{
global(name) <- value;

}

