% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{global}
\alias{global}
\title{Set variable in global}
\usage{
global(name)
}
\arguments{
\item{name}{The variable name string that going to
get its value in global environment.}
}
\value{
\code{NULL}

Get variable in global environment

The variable value of the given variable name. If the variable is
        not exists in global environment, then value \code{NULL} will be
        return.
}
\description{
Set variables into the global environment.
}
\examples{
global(name, value);

}
