% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{Imports}
\alias{Imports}
\title{VisualBasic namespace imports helper}
\usage{
Imports(namespace, overrides = FALSE, silent = TRUE,
  frame = parent.frame())
}
\arguments{
\item{namespace}{The namespace function name or function itself}

\item{overrides}{A logical flag to indicate that should the imported function
can overrides the previous function which they have the same
name. By default is can not. If this parameter is set to true,
then a warning message will be generated for mention you that
which functions are overrided.}

\item{frame}{The envrionment location, by default is current function scope}
}
\value{
A string vector contains all of the function names from this namespace
        function that imported to current environment.
}
\description{
VisualBasic namespace imports helper
}

