% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{include}
\alias{include}
\title{è¿™ä¸ªå‡½æ•°æ˜¯ä¸ºäº†å…¼å®¹è„šæœ¬æ¨¡å¼å’ŒRåŒ…æ¨¡å¼ä¸‹çš„å¼•ç”¨é—®é¢<U+FFFD>}
\usage{
include(R)
}
\description{
è¿™ä¸ªå‡½æ•°æ˜¯ä¸ºäº†å…¼å®¹è„šæœ¬æ¨¡å¼å’ŒRåŒ…æ¨¡å¼ä¸‹çš„å¼•ç”¨é—®é¢<U+FFFD>
}

