% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{is.nothing}
\alias{is.nothing}
\title{Determine that target is Nothing in VB way}
\usage{
is.nothing(...)
}
\description{
Determine that target is Nothing in VB way
}
\details{
This function is an alias of the \code{\link{IsNothing}} function.
}
\seealso{
\code{\link{IsNothing}}
}

