% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{IsNothing}
\alias{IsNothing}
\title{Determine that target \code{is.null} or \code{is.na} or \code{length}
equals to ZERO?}
\usage{
IsNothing(x.)
}
\arguments{
\item{x.}{Object with any type for determine that its value is null
or empty or not?}
}
\description{
Determine that target \code{is.null} or \code{is.na} or \code{length}
equals to ZERO?
}
\details{
åˆ¤æ–­å¯¹è±¡æ˜¯å¦ä¸ºç©ºï¼Œåœ¨è¿™ä¸ªå‡½æ•°é‡Œé¢ï¼Œç©ºå€¼ï¼ŒNAå€¼ï¼Œé•¿åº¦ä¸ºé›¶çš„å‘é‡ï¼Œ
         åˆ—è¡¨ç­‰éƒ½ä¼šè¢«å½“ä½œä¸ºç©ºå€<U+FFFD>
}

