% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{levenshtein.distance}
\alias{levenshtein.distance}
\title{Compute Levenshtein distance between two strings}
\usage{
levenshtein.distance(source, target, type = c("distance", "matrix"),
  insert.fun = function(x) 1, delete.fun = function(x) 1,
  substitute.fun = function(s, t) ifelse(s == t, 0, 1))
}
\arguments{
\item{source}{Source string.}

\item{target}{Target string.}

\item{type}{Specifies the return type. 'distance' for a single
distance value; 'matrix' for the matrix used during
dynamic programming.}

\item{insert.fun}{delete.fun, substitute.fun: Penalty functions of insert,
delete and substitute operation, whose return value must
be a single scalar value.}
}
\description{
Compute Levenshtein distance between two strings
}

