% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{log.close}
\alias{log.close}
\title{Close current log file}
\usage{
\method{log}{close}(print.warnings = FALSE)
}
\arguments{
\item{print.warnings}{Print all of the warnings before close
the log writer? Default is not print warnings.}
}
\description{
Close current log file
}

