% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{log.echo}
\alias{log.echo}
\title{Print a log text onto screen}
\usage{
\method{log}{echo}(...)
}
\arguments{
\item{...}{Parameters for \code{sprintf} function.}
}
\description{
Print a log text onto screen
}
