% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{log.open}
\alias{log.open}
\title{R logging helper by \code{sink}}
\usage{
\method{log}{open}(file.path)
}
\arguments{
\item{file.path}{The \code{*.log} log file path.}
}
\description{
This function can create the parent dir for the given
   log file safely.
}
