% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{LogException}
\alias{LogException}
\title{Logging error log file.}
\usage{
LogException(ex, logFile, append = FALSE)
}
\arguments{
\item{ex}{Can be error exception or error string text}

\item{logFile}{File path for save the log data}

\item{append}{a logical flag that indicate clear the content
before write the log file or not.}
}
\value{
Nothing
}
\description{
Logging error log file.
}
