% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{Matrix.XML}
\alias{Matrix.XML}
\title{Write the matrix/dataframe as XML node}
\usage{
Matrix.XML(matrix, indent, write, node.name = NULL)
}
\description{
Write the variable object of \code{matrix}/\code{data.frame} type
as a node in XML document.
}

