% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{memory}
\alias{memory}
\title{Get system memory info}
\usage{
memory()
}
\value{
A list contains two slot: \code{Mem} for installed physical memory 
   and \code{Swap} for swap memory.
   All of the memory data that returns from this function is in byte size 
   unit \code{MB}.
}
\description{
Get system memory info
}

