% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskHelper.R
\name{memory.size.auto}
\alias{memory.size.auto}
\title{Unify method for get memory size}
\usage{
memory.size.auto()
}
\description{
Provides a unify method for get current R session its
memory usage. For solve the \code{Inf} bug of \code{memory.size}
function on linux platform.
}
