% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linq.R
\name{Microsoft.VisualBasic.Data.Linq}
\alias{Microsoft.VisualBasic.Data.Linq}
\title{Linq helper in R language.}
\usage{
Microsoft.VisualBasic.Data.Linq()
}
\arguments{
\item{enumerable}{The data source collection for the group operation,
by default is ``data.frame`` type.}

\item{key}{The column name string (or a column projector function in data.frame)
or property name (in list) for using its corresponding value as the group
key}

\item{type}{Indicate that the data source ``enumerable`` is a data.frame or list?
       default is a data.frame.

Groups the items in a list by s specific property of the list item.}

\item{list}{The data source in list type}

\item{key}{The item property name in this list source collection.

Groups the rows in a dataframe by a specific column as key.}

\item{data.frame}{The data source in data.frame type}

\item{key}{The column name in string mode or column projector function
    for read the column data in target source as the group key.
Group the string collection
Example:

\code{
  Group(c("A","B","C","c")) %=>% Count;\cr\cr

  $A: 1\cr
  $B: 1\cr
  $C: 2\cr
}

This function is usually piped with the \code{Group} function}
}
\value{
\enumerate{
    \item \code{WhichIsNotEmpty} returns the index in the input
          \code{vector}/\code{list} sequence where is not nothing.
}
group data.frame/list by keys
}
\description{
Linq helper in R language.
}
