% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{MyList}
\alias{MyList}
\title{List all VisualBasic namespace}
\usage{
MyList()
}
\value{
Returns a function name character vector which its name is stared by token like:
    \code{Microsoft.VisualBasic}.
}
\description{
List all VisualBasic namespace
}
