% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{node.find}
\alias{node.find}
\title{Find a node value by a given search key}
\usage{
node.find(tree, search, key.compares)
}
\arguments{
\item{search}{Object key for search a tree node.}
}
\description{
Find a node value by a given search key
}
