% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{node.right}
\alias{node.right}
\title{Get the right node of current node}
\usage{
node.right(tree, node)
}
\arguments{
\item{tree}{The binary tree list object.}

\item{node}{The current tree node}
}
\description{
Get the right node of current node
}
