% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{numeric.group.impl}
\alias{numeric.group.impl}
\title{Group a numeric vector without check}
\usage{
numeric.group.impl(seq, assert)
}
\arguments{
\item{seq}{Please ensure that this numeric sequence is not
\code{NULL} or \code{NA} value, and it must contains
more than 1 elements.}
}
\description{
Group a numeric vector without check
}

