% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearCluster.R
\name{numeric.group}
\alias{numeric.group}
\title{Group a numeric vector}
\usage{
numeric.group(seq, assert = function(x, y) abs(x - y) <= 1)
}
\arguments{
\item{seq}{A numeric sequence}

\item{assert}{A given test condition for test if a number is a member
of the current group or not?}
}
\description{
Group a numeric vector elements by a given test condition
}
