% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{package.is_missing}
\alias{package.is_missing}
\title{Determine a given package is missing?}
\usage{
package.is_missing(package)
}
\arguments{
\item{package}{A package name string}
}
\value{
A logical value to indicate that the given package is missing from
   current environment or not?
}
\description{
Determine a given package is missing?
}

