% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{package.missing}
\alias{package.missing}
\title{Determine the missing packages}
\usage{
package.missing(package)
}
\arguments{
\item{package}{A character vector of the package names.}
}
\value{
Function returns the name list of the missing packages.
}
\description{
Determine the missing packages
}

