% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{package.version}
\alias{package.version}
\title{Get version string for package}
\usage{
package.version(package = "VisualBasic.R")
}
\arguments{
\item{package}{The package name}
}
\value{
Function returns the package version string value. If the package is not exists,
   then this function returns \code{NA} value.
}
\description{
Due to the reason of the systen function \code{package_version} using
   integer value for represent the package version string. So this may cause bugs.
   This function using string for represent the package version to avoid the bugs.
}

