% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{parse.package.description}
\alias{parse.package.description}
\title{Parse DESCRIPTION file}
\usage{
parse.package.description(
  file = base::system.file("DESCRIPTION", package = "VisualBasic.R")
)
}
\arguments{
\item{file}{\code{DESCRIPTION} file handle}
}
\value{
Returns a list of named value that parsed from the \code{DESCRIPTION}
     file in a specific package.
}
\description{
Parse DESCRIPTION file
}
