% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{primitiveTypes}
\alias{primitiveTypes}
\title{æšä¸¾å‡ºRè¯­è¨€ä¹‹ä¸­çš„ä¸€äº›åŸºç¡€çš„æ•°æ®ç±»å‹ï¼Œæšä¸¾å€¼æœ‰ï¼<U+FFFD>
\enumerate{
\item \code{object} = 0
\item \code{data.frame} = 1
\item \code{list} = 2
\item \code{vector} = 3
}}
\usage{
primitiveTypes()
}
\description{
æšä¸¾å‡ºRè¯­è¨€ä¹‹ä¸­çš„ä¸€äº›åŸºç¡€çš„æ•°æ®ç±»å‹ï¼Œæšä¸¾å€¼æœ‰ï¼<U+FFFD>
\enumerate{
\item \code{object} = 0
\item \code{data.frame} = 1
\item \code{list} = 2
\item \code{vector} = 3
}
}

