% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{printf}
\alias{printf}
\title{Simulate the C \code{printf} function}
\usage{
printf(...)
}
\arguments{
\item{...}{Function parameter for function \code{\link{sprintf}}}
}
\description{
Simulate the C \code{printf} function
}

