% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{Println}
\alias{Println}
\title{Append text content to file}
\usage{
Println(file.txt, content)
}
\arguments{
\item{file.txt}{File path}

\item{content}{The text content that will be write to
target text file.
It can be a string lines vector.}
}
\value{
Nothing
}
\description{
Append text content to a specific text file.
}

