% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{push.x}
\alias{push.x}
\title{Choose different XML write function based on the input object type.}
\usage{
push.x(x, indent, write, name = NULL)
}
\description{
Choose different XML write function based on the input object type.
}

