% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{ReadAllText}
\alias{ReadAllText}
\title{Read all text from a specific file}
\usage{
ReadAllText(file.txt)
}
\arguments{
\item{file.txt}{The target file path for read}
}
\value{
Returns the text file content in one piece, 
   not split in lines.
}
\description{
Read all text from a specific file
}
