% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{SaveXML}
\alias{SaveXML}
\title{Serialize any R object to a specific XML file, for save data in a more
human readable format.}
\usage{
SaveXML(x, file.xml, rootName = "Rlang.xml")
}
\arguments{
\item{x}{The R object in any type}

\item{file.xml}{The file path of the XML file dump data will be saved.}

\item{rootName}{The name of the generated xml root node.}
}
\description{
Serialize any R object to a specific XML file, for save data in a more
human readable format.
}

