% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{Size}
\alias{Size}
\title{Returns the object size}
\usage{
Size(x)
}
\arguments{
\item{x}{R object in any type}
}
\value{
A list with element: \code{rows} and \code{cols} to indicate the object size.\cr\cr

        for x is \code{data.frame}, these two element value will be \code{\link{base::nrows}} and \code{\link{base::ncols}}\cr
        for x is \code{list} or \code{vector}, these two element value will be \code{rows = 1} and \code{cols = \link{length}(x)}\cr
        for x is object like S4 class, these two element value will be \code{[1,1]}\cr
}
\description{
Returns the object size
}

