% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slave.R
\name{slave_closure}
\alias{slave_closure}
\title{Run a closure string}
\usage{
slave_closure(closure = "stdin")
}
\arguments{
\item{closure}{A string vector that can represent as a function, which should have no parameter.
By default, the closure text is read from stdin.}
}
\description{
The closure string could be content from stdin or user input text.
}
