% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidewindow.R
\name{slide.windows}
\alias{slide.windows}
\title{Create slide windows}
\usage{
slide.windows(win_size = 2, step = 1, ...)
}
\arguments{
\item{...}{A list of named variable which should be vectors and theirs vector size
should also be equals to each other.}
}
\description{
Create slide windows
}
