% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{sort.dataframe}
\alias{sort.dataframe}
\title{å¯¹æ•°æ®æ¡†æŒ‰ç…§æŒ‡å®šçš„åˆ—è¿›è¡Œæ’åº}
\usage{
\method{sort}{dataframe}(dataframe, key, key.numeric = function(v)
  as.numeric(v), desc = FALSE)
}
\arguments{
\item{key}{æ•°æ®æ¡†ä¹‹ä¸­çš„åˆ—åç§°æˆ–è€…åˆ—çš„ç´¢å¼•ç¼–å<U+FFFD>}

\item{key.numeric}{è¿™ä¸ªlambdaå‡½æ•°æè¿°äº†å¦‚ä½•å°†æ‰€æŒ‡å®šçš„åˆ—çš„å€¼è½¬æ¢ä¸ºç”¨æ¥è¿›è¡Œæ’åºæ‰€éœ€è¦çš„æ•°å€¼ä¾æ®çš„è¿‡ç¨‹}

\item{desc}{æ˜¯å¦æ‰§è¡Œé™åºæ’åºï¼Ÿé»˜è®¤æ˜¯å‡åºæ’åº}
}
\description{
å¯¹æ•°æ®æ¡†æŒ‰ç…§æŒ‡å®šçš„åˆ—è¿›è¡Œæ’åº
}

