% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{sort.list}
\alias{sort.list}
\title{æˆ‘ä»¬å‡è®¾åœ¨è¿™é‡Œçš„listæ˜¯é‡Œé¢çš„æ‰€æœ‰çš„å…ƒç´ éƒ½æ˜¯listå¯¹è±¡ï¼Œå¹¶ä¸”å…ƒç´ çš„åç§°éƒ½ç›¸å<U+FFFD>}
\usage{
\method{sort}{list}(list, key, key.numeric = function(v) as.numeric(v),
  desc = FALSE)
}
\arguments{
\item{list}{The input list object}

\item{key}{A lambda function or property name string for get key value for the sort operation.}

\item{key.numeric}{A lambda function for evaluate the key value to numeric value}

\item{desc}{A logical flag to indicated that sort the input sequence in ASC or DESC mode?}
}
\value{
A data sequence which its elements has been reordered.
}
\description{
æˆ‘ä»¬å‡è®¾åœ¨è¿™é‡Œçš„listæ˜¯é‡Œé¢çš„æ‰€æœ‰çš„å…ƒç´ éƒ½æ˜¯listå¯¹è±¡ï¼Œå¹¶ä¸”å…ƒç´ çš„åç§°éƒ½ç›¸å<U+FFFD>
}

