% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.Empty}
\alias{Strings.Empty}
\title{Determine that target string is null or empty or not??}
\usage{
Strings.Empty(s, NA.empty = FALSE)
}
\description{
Determine that target string is null or empty or not??
}

