% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.Join}
\alias{Strings.Join}
\title{Contact string tokens}
\usage{
Strings.Join(SourceArray, Delimiter = " ")
}
\arguments{
\item{SourceArray}{Required. One-dimensional array containing substrings to be joined.}

\item{Delimiter}{Optional. Any string, used to separate the substrings in the
 returned string. If omitted, the space character (" ") is used. If
Delimiter is a zero-length string ("") or Nothing, all items in the
 list are concatenated with no delimiters.}
}
\value{
Returns a string created by joining a number of substrings contained in an
     array.
}
\description{
Returns a string created by joining a number of substrings
             contained in an array.
}

