% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.LCase}
\alias{Strings.LCase}
\title{String to lowercase}
\usage{
Strings.LCase(Value)
}
\arguments{
\item{Value}{Required. Any valid String or Char expression.}
}
\value{
Returns a string or character converted to lowercase.
}
\description{
Returns a string or character converted to lowercase.
}

