% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.Replace}
\alias{Strings.Replace}
\title{Text replacement}
\usage{
Strings.Replace(Expression, Find, Replacement)
}
\arguments{
\item{Expression}{Required. String expression containing substring to replace.}

\item{Find}{Required. Substring being searched for.}

\item{Replacement}{Required. Replacement substring.}

\item{Start}{Optional. Position within Expression that starts a substring
used for replacement. The return value of Replace is a string that begins
at Start, with appropriate substitutions. If omitted, 1 is assumed.}

\item{Count}{Optional. Number of substring substitutions to perform.
If omitted, the default value is -1, which means "make all possible
substitutions."}

\item{Compare}{Optional. Numeric value indicating the kind of comparison
to use when evaluating substrings. See Settings for values.}
}
\value{
Replace returns the following values.IfReplace returns Find
       is zero-length or NothingCopy of Expression Replace is zero-length
       Copy of Expression with no occurrences of Find Expression is zero-length
       or Nothing, or Start is greater than length of Expression Nothing Count is 0
       Copy of Expression
}
\description{
Returns a string in which a specified substring has been
         replaced with another substring a specified number of times.
}

