% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.Split}
\alias{Strings.Split}
\title{Split string using regexp}
\usage{
Strings.Split(Expression, Delimiter = " ", Compare = 0)
}
\arguments{
\item{Expression}{Required. String expression containing substrings and delimiters.}

\item{Delimiter}{Optional. Any single character used to identify substring limits.
If Delimiter is omitted, the space character (\code{" "}) is assumed to be
the delimiter.}

\item{Compare}{Optional. Numeric value indicating the comparison to use when
evaluating substrings. See "Settings" for values.}

\item{Limit}{Optional. Maximum number of substrings into which the input string
should be split. The default, ?, indicates that the input string should
be split at every occurrence of the Delimiter string.}
}
\value{
String array. If Expression is a zero-length string (\code{""}), Split returns
       a single-element array containing a zero-length string. If Delimiter is a
       zero-length string, or if it does not appear anywhere in Expression, Split
       returns a single-element array containing the entire Expression string.
}
\description{
Returns a zero-based, one-dimensional array containing a specified
             number of substrings.
}
