% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.UCase}
\alias{Strings.UCase}
\title{String in uppercase}
\usage{
Strings.UCase(Value)
}
\arguments{
\item{Value}{Required. Any valid String or Char expression.}
}
\value{
Returns a string or character containing the specified string
        converted to uppercase.
}
\description{
Returns a string or character containing the specified string
        converted to uppercase.
}

