% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskHelper.R
\name{stripREnvironmentInfo}
\alias{stripREnvironmentInfo}
\title{Removes R environment info}
\usage{
stripREnvironmentInfo(closure)
}
\arguments{
\item{closure}{A function tostring result.}
}
\description{
A function is convert to string by \code{capture.output} function
   may contains some R environment information like:
   \code{<bytecode: 0x1872cbb0><environment: 0x241cc710>}. The information could make
   the \code{eval(parse(text = closure))} process failure. This helper tools
   will try to removes such environment information text.
}

