% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{substr.Right}
\alias{substr.Right}
\title{Substring from right}
\usage{
substr.Right(x, n)
}
\arguments{
\item{x}{A given string vector}

\item{n}{Length of the substring}
}
\description{
This function working in the reverse direction of function
     \code{\link{Mid}}
}
