% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{swap}
\alias{swap}
\title{Swap two variable}
\usage{
swap(a, b)
}
\arguments{
\item{a}{variable 1}

\item{b}{variable 2}
}
\description{
Swap two variable
}
\details{
example as:

   a <- 1;
   b <- "Hello World!";
   c(b, a) := swap(a, b);
}
