% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextWriter.R
\name{textWriter}
\alias{textWriter}
\title{Open a text writer}
\usage{
textWriter(path, buffer.Size = 8192)
}
\arguments{
\item{path}{The file path string.}

\item{buffer.Size}{Buffer size in number of text lines.}
}
\value{
A text writer handler function.
}
\description{
Open a text writer
}

