% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{tick.each}
\alias{tick.each}
\title{Display progress for elements' action}
\usage{
tick.each(sequence, action)
}
\arguments{
\item{sequence}{A list or vector object.}

\item{action}{A function that will applied on each element
in the input sequence.}
}
\value{
action result on each elements
}
\description{
Only works on the \code{vector} or \code{list}
}
\details{
\code{\link{tick.helper}}
}

