% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{tick.helper}
\alias{tick.helper}
\title{Helper function for display the progress in percentage value of a
task which is required very long time to run.}
\usage{
tick.helper(total, step = 5/100)
}
\arguments{
\item{total}{The total elements in the Long time run task that will be processing}
}
\value{
A lambda function that using for report the task progress.
}
\description{
Only avaliable for single thread R app
}

