% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Trim}
\alias{Trim}
\title{Returns string w/o leading or trailing whitespace}
\usage{
Trim(str)
}
\description{
https://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r
}

