% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{unix.timestamp}
\alias{unix.timestamp}
\title{Current unix timestamp in milliseconds}
\usage{
unix.timestamp()
}
\value{
timestamp in milliseconds
}
\description{
https://stackoverflow.com/questions/40059573/r-get-current-time-in-milliseconds
}

