% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{using}
\alias{using}
\title{Auto dispose}
\usage{
using(proc)
}
\arguments{
\item{proc}{This proc object should contains
a \code{dispose} function for release resource and
a \code{invoke} function to run your code.}
}
\description{
Auto dispose
}
