% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{vector.fill}
\alias{vector.fill}
\title{fill data vector}
\usage{
vector.fill(list, baselist)
}
\arguments{
\item{list}{Target value vector}

\item{baselist}{A vector provide length or a vector length integer number.}
}
\value{
A vector that with equals length with \code{baselist} vector.
}
\description{
fill data vector
}

