% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{xLoad}
\alias{xLoad}
\title{Load \code{rda} data in a unified method}
\usage{
xLoad(rdaName, envir = globalenv(), verbose = FALSE)
}
\arguments{
\item{rda}{The \code{*.rda} file path or package data name}
}
\description{
Load \code{rda} data in a unified method
}

