% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{Xml.Write.Any}
\alias{Xml.Write.Any}
\title{Write R object to Xml}
\usage{
Xml.Write.Any(x, indent, write, name = NULL)
}
\arguments{
\item{x}{Type is limited to \code{vector}, \code{list} and
\code{data.frame}/\code{matrix}.}

\item{indent}{The xml node indent, whitespace numbers.}

\item{write}{Text file write handle, which is created from \code{\link{File.Open}}}

\item{name}{Using this function parameter to overrides the node name.}
}
\description{
Choose different XML write function based on the input object
   type.
}
