% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{Xml.Write.List}
\alias{Xml.Write.List}
\title{Write the \code{list} as XML node}
\usage{
Xml.Write.List(list, indent, write, node.name = NULL)
}
\description{
Write the \code{list} as XML node
}
\details{
A R list object is a kind of dictionary object in VB.NET
}
