% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{argv}
\alias{argv}
\title{Get commandline}
\usage{
argv(arguments = base::commandArgs())
}
\description{
Get commandline
}
\note{
Commandline parser for cli expression pattern like:

\code{Rscript script.R /name /arg1 value1 /arg2 value2 /boolean1 /arg3 value3}

The \code{nextToken} function in the returns list object is
usually using for get value by a given key name.
}
