% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{as.index}
\alias{as.index}
\title{Key index helper}
\usage{
as.index(keys)
}
\arguments{
\item{keys}{This function parameter should be a string character vector.
If this parameter its value is null, then this function will returns all
of the keys that in target index closure.}
}
\value{
If the given \code{keys} is NULL or a single NA value, 
        then NULL value will be returns. Otherwise, returns a 
        index function.
}
\description{
Create a index object for determine that a given id
is exists in the keys collection or not.
}
