% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{as.index}
\alias{as.index}
\title{Key index helper}
\usage{
as.index(keys)
}
\arguments{
\item{keys}{This function parameter should be a string character vector.}
}
\description{
Create a index object for determine that a given id 
is exists in the keys collection or not.
}

