% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{binarySearch}
\alias{binarySearch}
\title{å¯¹ä¸€ä¸ªå·²ç»è¿›è¡Œæ’åºçš„åºåˆ—è¿›è¡ŒäºŒåˆ†æ³•æŸ¥æ‰<U+FFFD>}
\usage{
binarySearch(source, find, key, compares = function(a, b) a - b)
}
\arguments{
\item{source}{å¿…é¡»è¦å¯¹æŸ¥æ‰¾çš„é”®è¿›è¡Œäº†å‡åºæ’åºæ“ä½<U+FFFD>}

\item{find}{æŸ¥æ‰¾çš„ç›®æ ‡key}

\item{key}{ç”¨äºæè¿°å¦‚ä½•ä»sourceä¹‹ä¸­çš„ä¸€ä¸ªå…ƒç´ å¾—åˆ°è¿›è¡Œæ¯”è¾ƒçš„keyçš„lambdaè¡¨è¾¾å¼æ–¹æ³<U+FFFD>}

\item{compares}{findå¾…æŸ¥æ‰¾ç›®æ ‡é”®keyå’Œsourceä¹‹ä¸­çš„å¯¹è±¡çš„keyçš„æ¯”è¾ƒæ–¹å¼ï¼Œé»˜è®¤æ˜¯æŒ‰ç…§æ•°å€¼å¤§å°è¿›è¡Œæ¯”è¾<U+FFFD>}
}
\description{
å¯¹ä¸€ä¸ªå·²ç»è¿›è¡Œæ’åºçš„åºåˆ—è¿›è¡ŒäºŒåˆ†æ³•æŸ¥æ‰<U+FFFD>
}

