% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{binaryTree}
\alias{binaryTree}
\title{Build a binary tree list}
\usage{
binaryTree(src, key, key.compares, debug = FALSE, progressHeader = TRUE)
}
\arguments{
\item{src}{the input data sequence, which can be \code{data.frame}, \code{list} or
\code{vector}}

\item{key}{An object indexer for get the key value for the elements}

\item{key.compares}{A function to compare two key, this function
    should returns a integer value:

    \enumerate{
       \item \code{0} The two tree node key is equals to each other.
       \item \code{1} The \code{key1} is greater than \code{key2}.
       \item \code{-1} The \code{key1} is less than \code{key2}.
    }}
}
\value{
A binary tree S4 class object
}
\description{
This function build a binary tree list.
}

