% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enumerator.R
\name{enumerator}
\alias{enumerator}
\title{Enumerator object constructor function for create a VB.NET linq like data sequence wrapper.
This constructor function will create a list which contains a sequence source and sevral
linq data helper extension functions}
\usage{
enumerator(src)
}
\arguments{
\item{src}{A generic type data sequence, which can be a \code{dataframe}, \code{list}, or \code{vector}.}
}
\value{
A generic enumerator list object, which it contains:
        \enumerate{
        \item \code{src} The generic data sequence input
        \item \code{where} Extension function for select values by test on condition for each element if they are true
        \item \code{select} Project the input data sequence source to another form.
        \item \code{toarray} Get the input source data sequence.
        }
}
\description{
Enumerator object constructor function for create a VB.NET linq like data sequence wrapper.
This constructor function will create a list which contains a sequence source and sevral
linq data helper extension functions
}

