% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{GetType}
\alias{GetType}
\title{This function returns on of the member value from enumeration function
\code{\link{primitiveTypes}}
based on the type of the variable \code{x}}
\usage{
GetType(x)
}
\arguments{
\item{x}{Any variable}
}
\value{
One of the value member from enumeration function
        \code{\link{primitiveTypes}}.
}
\description{
This function returns on of the member value from enumeration function
\code{\link{primitiveTypes}}
based on the type of the variable \code{x}
}

