% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{include}
\alias{include}
\title{This function is using for solving the code compatibility
between the R package and R script.}
\usage{
include(R)
}
\arguments{
\item{R}{R script file path.}
}
\value{
Nothing
}
\description{
This function is using for solving the code compatibility
between the R package and R script.
}
