% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{IsNothing}
\alias{IsNothing}
\title{Determine that target is Nothing in VB way}
\usage{
IsNothing(x, stringAsFactor = FALSE)
}
\arguments{
\item{x}{Object with any type for determine that its value is null
or empty or not?}

\item{stringAsFactor}{A logical flag that indicated that should treated
the string value like \code{NULL}, \code{NA}, etc
as Nothing?}
}
\description{
Determine that target is Nothing in VB way.
             \code{is.null} or \code{is.na} or \code{length} equals to ZERO?
             All of these situation will be treated as Nothing.
}
\details{
Determine that target object is nothing or not, by using predicates: \code{is.null},
   \code{is.na}, \code{length(x) == 0}. And if the function parameter \code{stringAsFactor} is
   true, then string comparision for \code{x == ""}, \code{x == "NULL"}, \code{x == "NA"} will
   be applied.
}
\seealso{
This function has an alias name: \code{\link{is.nothing}}.
}
