% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{microsoft.visualbasic.data}
\alias{microsoft.visualbasic.data}
\title{Namespace that contains for some common data operation helper function}
\usage{
microsoft.visualbasic.data()
}
\description{
Namespace that contains for some common data operation helper function
}

