% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Microsoft.VisualBasic.Data}
\alias{Microsoft.VisualBasic.Data}
\title{Common data operation helper function}
\usage{
Microsoft.VisualBasic.Data()
}
\description{
Common data operation helper function
}
\details{
\enumerate{
   \item \code{ensure.dataframe} Fix for a bug in R language: for ensure that
         when get one row from a dataframe didn't output a list object, this
         function make sure the subset of a dataframe is always output a
         dataframe object.
   \item \code{.as.list} Converts the \code{data.frame} object to a set of
         \code{list} object from each rows data.
}
}

