% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Microsoft.VisualBasic.Data}
\alias{Microsoft.VisualBasic.Data}
\title{Common data operation helper function}
\usage{
Microsoft.VisualBasic.Data()
}
\arguments{
\item{project}{Column names subset for create a dataframe with selected fields.

Convert a list object as dataframe}

\item{list}{A list object, with members is vector

ensure the result is a dataframe object

There is a bug in R dataframe subset operation:
If the subset result just one row,
Then the subset result will be convert a vector automatically,
not a dataframe any more. This will cause bugs and code inconsist.
Using this function to ensure all of your dataframe subset result
is a dataframe object, not a vector.}
}
\description{
Common data operation helper function
}
\details{
\enumerate{
   \item \code{ensure.dataframe} Fix for a bug in R language: for ensure that
         when get one row from a dataframe didn't output a list object, this
         function make sure the subset of a dataframe is always output a
         dataframe object.

   \item \code{.as.list} Converts the \code{data.frame} object to a set of
         \code{list} object from each rows data.

   \item \code{read.dataset} Read the csv file as matrix, the first column from
         the csv file will be using as the rownames and the first row will be
         using as the colnames.
}
Convert a list of list object as dataframe
}

