% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Language.R
\name{Microsoft.VisualBasic.Language}
\alias{Microsoft.VisualBasic.Language}
\title{Syntax tweaks helpers}
\usage{
Microsoft.VisualBasic.Language()
}
\arguments{
\item{x}{Function parameter}

\item{y}{Function with one parameter}

\item{x}{vectors (of the same mode) containing a sequence
of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
of items (conceptually) with no duplicated values.}
}
\value{
Namespace module with exports:
    \enumerate{
       \item \code{\%||\%} The \code{OR} default value operator
       \item \code{\%:=\%} The tuple syntax
       \item \code{\%=>\%} Forward pipeline operator
       \item \code{\%+\%}  Set union operator
    }
The \code{OR} default expression

\code{\%||\%}: The target given value or default value is
    target value \code{\link{IsNothing}}.
Extension method operator

\code{\%=>\%}: The function value.
Union two collection

\code{\%+\%}: A vector of the same mode as x or y for a common mode for \code{union}.
}
\description{
Syntax tweaks helpers
}

