% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Mid}
\alias{Mid}
\title{Substring from left}
\usage{
Mid(s, start, length)
}
\arguments{
\item{s}{A given string vector}

\item{start}{The substring start location on \code{s}}

\item{length}{The length of the substring}
}
\description{
Substring from left
}

