% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{primitiveTypes}
\alias{primitiveTypes}
\title{Enumerate some primitive type in R language, these enumeration value have:}
\usage{
primitiveTypes()
}
\description{
\enumerate{
\item \code{object} = 0
\item \code{data.frame} = 1
\item \code{list} = 2
\item \code{vector} = 3
}
}

