% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskHelper.R
\name{slave}
\alias{slave}
\title{Run closure in a new R process}
\usage{
slave(closure)
}
\arguments{
\item{closure}{should be a function without any parameter. And have no function returns.}
}
\description{
This is usually using for memory size optimization.
}

