% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskHelper.R
\name{slave}
\alias{slave}
\title{Run closure in a new R process}
\usage{
slave(closure, arguments = NULL, parallel = FALSE, debug = FALSE)
}
\arguments{
\item{closure}{should be a function without any parameter. And have no function returns.}

\item{arguments}{The additional commandline arguments that will pass to the child process.
This parameter should be a named list or character vector. If the argument list object
contains a \code{commandName} member, then it will generates a VisualBasic style
commandline arguments, which can be processed by \code{\link{argv}} function.}

\item{parallel}{This not parallel, the slave closure function calls will block running of
R program at here. By default is not running in parallel mode.}
}
\description{
This is usually using for memory size optimization.
}

