% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarySearch.R
\name{sort.dataframe}
\alias{sort.dataframe}
\title{Sort a dataframe by a specific given column name or key indexer.}
\usage{
\method{sort}{dataframe}(dataframe, key, key.numeric = function(v)
  as.numeric(v), desc = FALSE)
}
\arguments{
\item{key}{The key column name or a lambda function to summary the
rows to a specific key}

\item{key.numeric}{This lambda function describ how to converts the specific key
column/indexer values to the numeric values which is use for
sort the rows based on these numeric value comparsion.}

\item{desc}{Sort in descending order? by default is not.}
}
\description{
Sort a dataframe by a specific given column name or key indexer.
}

