% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\name{Strings.Empty}
\alias{Strings.Empty}
\title{A given string is empty?}
\usage{
Strings.Empty(s, NA.empty = FALSE)
}
\arguments{
\item{s}{A given string vector that using for measurement}

\item{NA.empty}{This logical flag indicate that should treat the text which is
equals to \code{chr("NA")} as empty too? Default is not.}
}
\value{
Logical vector, if the input parameter s is NULL, then this function 
     will also returns a logical value TRUE.
}
\description{
Determine that target string is null or empty or not??
}

