% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskHelper.R
\name{tick.helper}
\alias{tick.helper}
\title{Progress tick on console}
\usage{
tick.helper(total, disp.number = TRUE, step = 5/100)
}
\arguments{
\item{total}{The total elements in the Long time run task that will be processing}

\item{step}{The tick interval value for report the task progress value, default is \code{5/100}.}
}
\value{
A lambda function that using for report the task progress.
}
\description{
Helper function for display the progress in percentage value of a
             task which is required very long time to run.

             Only avaliable for single thread R app
}

