% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{xLoad}
\alias{xLoad}
\title{Load \code{rda} data in a unified method}
\usage{
xLoad(rdaName, envir = globalenv(), verbose = FALSE)
}
\arguments{
\item{rdaName}{The \code{*.rda} file path or package data name.
If this parameter is empty string, then load function will
throw error;
If the rda data object could not be found, this function just givens warnings.}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
Load \code{rda} data in a unified method
}
\details{
The file path string value of the \code{rdaName} could be
a relative path or an absolute path.
When the file path is a relative path, then the R script will try to
search the given rda file in directories:

\enumerate{
  \item \code{.} Current workspace.
  \item \code{./data} The \code{data} folder in the current workspace.
  \item \code{../data} The \code{data} folder in current workspace's parent directory.
}
}
