% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{XML.Framework}
\alias{XML.Framework}
\title{A framework for write R object to XML file.}
\usage{
XML.Framework(write, do.write, rootName)
}
\arguments{
\item{write}{The file write handle from the \code{\link{File.Open}} or \code{\link{textWriter}} function.
It is recommended use \code{\link{textWriter}} function for better performance.}

\item{do.write}{A function pointer that used for describ how to build the output xml file}

\item{rootName}{The node name of the generated xml root node.}
}
\description{
A framework for write R object to XML file.
}
