% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{Xml.Write.Matrix}
\alias{Xml.Write.Matrix}
\title{Write the matrix/dataframe as XML node}
\usage{
Xml.Write.Matrix(matrix, indent, write, node.name = NULL)
}
\description{
Write the variable object of \code{matrix}/\code{data.frame}
  type as a node in XML document.
}
\details{
Output the xml node in format like:

\code{

   <node.name nrows = ...>\cr
      <tr rowname = ...>\cr
         <td colname = >value</td>\cr
      </tr>\cr
      <tr>\cr
      </tr>\cr
   </node.name>

}
}

