% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML.R
\name{Xml.Write.Vector}
\alias{Xml.Write.Vector}
\title{Write the vector as the xml node}
\usage{
Xml.Write.Vector(vector, indent, write, name = NULL)
}
\description{
Write the vector as the xml node:

  If the vector type is characters, then this function will generates the
  resulted xml in a list format;

  If the vector type is numeric or logical, then the vector will be saved
  as xml attribute.
}

